requirements.txt
numpy==1.16.3
pandas>=0.22.0
six>=1.11.0
scipy>=0.19.1
scikit-learn>=0.22
Keras==2.1.5
tensorflow==1.3.0
Pillow==4.0.0
h5py==2.7.0
progressbar2
tabulate
matplotlib
gensim==3.4.0
# Add for experiment
featuretools>=0.13.2
auto-sklearn>=0.6.0